Minor fixes in handling of decrypt stdout stderr tempfiles
out is a tempfile.NamedTemporaryFile created with delete=True,
(default) so it'll get up deleted on disk once the file is closed.

err is a tempfile.NamedTemporaryFile created with delete=False,
so we'll need to delete it after closing it.

This commit just makes sure both err and out are deleted after use.